# backend/src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Create order",
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "description": "ID of the customer",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/lines": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Add product to order",
        "operationId": "addOrderLine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderLineDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Product added to order",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Product not found or Product is inactive or Order must not be FINISHED, CANCELED or READY to add or remove order line",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "order-controller"
        ],
        "summary": "Remove product from order",
        "operationId": "removeOrderLine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderLineDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Product removed from order",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Product not found or Product is inactive or Order must not be FINISHED, CANCELED or READY to add or remove order line",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{id}/prepare": {
      "patch": {
        "tags": [
          "order-controller"
        ],
        "summary": "Prepare order",
        "operationId": "prepareOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Order prepared",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/notify": {
      "patch": {
        "tags": [
          "order-controller"
        ],
        "summary": "Notify prepared order",
        "operationId": "notifyPreparedOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Order notified as prepared",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/finish": {
      "patch": {
        "tags": [
          "order-controller"
        ],
        "summary": "Finish order",
        "operationId": "finishOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Order finished",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/confirm": {
      "patch": {
        "tags": [
          "order-controller"
        ],
        "summary": "Confirm order",
        "operationId": "confirmOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order confirmed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/cancel": {
      "patch": {
        "tags": [
          "order-controller"
        ],
        "summary": "Cancel order",
        "operationId": "cancelOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Order canceled",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "Get order data",
        "operationId": "getOrderData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Order data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDTO"
                }
              }
            }
          }
        }
      }
    },
    "/orders/of-the-day": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "List orders of the day by state",
        "operationId": "listOrdersOfTheDayByState",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "State of the order",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "CREATED",
                "PENDING",
                "ACCEPTED",
                "IN_PREPARATION",
                "READY",
                "FINISHED",
                "CANCELED"
              ]
            },
            "examples": {
              "status": {
                "summary": "CANCELED",
                "description": "status",
                "value": "CANCELED"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders of the day by state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderLineDTO": {
        "required": [
          "category",
          "description",
          "name",
          "price",
          "productId",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "note": {
            "type": "string"
          }
        },
        "description": "Product to be removed"
      },
...
      "OrderDTO": {
        "required": [
          "id",
          "initialDateTime",
          "lines",
          "status",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineDTO"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING",
              "ACCEPTED",
              "IN_PREPARATION",
              "READY",
              "FINISHED",
              "CANCELED"
            ]
          },
          "total": {
            "type": "number"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "initialDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "finishedDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "trackOrderCode": {
            "type": "string"
          }
        }
      }
    }
  }
}